projects/19_secure_api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pyjwt==2.8.0
python-multipart==0.0.6
